# xiandeng-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "所有/v1/api下面的接口",
        "title": "研伴后台接口文档",
        "termsOfService": "http://xiandeng.net.cn",
        "contact": {
            "name": "Yancy Huang",
            "url": "http://xiandeng.net.cn",
            "email": "yhsunwest@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "http://localhost:8080",
    "paths": {
...
        "/system/log_common_message": {
            "post": {
                "description": "写普通日志",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统"
                ],
                "summary": "写普通日志",
                "parameters": [
                    {
                        "description": "BodyLogMessage",
                        "name": "BodyLogMessage",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/controller.BodyLogMessage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/system/log_message": {
            "post": {
                "description": "写日志",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统"
                ],
                "summary": "写日志",
                "parameters": [
                    {
                        "description": "BodyLogMessage",
                        "name": "BodyLogMessage",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/controller.BodyLogMessage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/system/order/proof/{order_id}": {
            "get": {
                "description": "列出订单凭证",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统"
                ],
                "summary": "列出订单凭证",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order_id",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "description": "上传订单凭证",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统"
                ],
                "summary": "UploadOrderProof",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order_id",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
...
        "/system/webhook": {
            "get": {
                "description": "回调接口",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统"
                ],
                "summary": "回调接口",
                "parameters": [
                    {
                        "type": "string",
                        "example": "echostr",
                        "description": "微信验证webhook接口时的字符串，需要返回",
                        "name": "echostr",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.BodyLogMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
...
}